Refactor classes into tracker package
