+++ src/tracker/ExpenseTrackerGUI.java
@@ -1,216 +1,12 @@
-package src;
+package tracker;
 
 import javax.swing.*;
 import javax.swing.table.DefaultTableModel;
 import java.awt.*;
-import java.sql.*;
-import java.text.ParseException;
 import java.text.SimpleDateFormat;
 import java.util.ArrayList;
 import java.util.Date;
 
-// Base Expense class
-abstract class Expense {
-    private double amount;
-    private Date date;
-    private String description;
-
-    public Expense(double amount, Date date, String description) {
-        this.amount = amount;
-        this.date = date;
-        this.description = description;
-    }
-
-    public double getAmount() {
-        return amount;
-    }
-
-    public Date getDate() {
-        return date;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public abstract String getType();
-}
-
-// FoodExpense class
-class FoodExpense extends Expense {
-    private String restaurant;
-
-    public FoodExpense(double amount, Date date, String description, String restaurant) {
-        super(amount, date, description);
-        this.restaurant = restaurant;
-    }
-
-    @Override
-    public String getType() {
-        return "Food";
-    }
-
-    public String getRestaurant() {
-        return restaurant;
-    }
-}
-
-// TravelExpense class
-class TravelExpense extends Expense {
-    private String destination;
-    private String transportMode;
-
-    public TravelExpense(double amount, Date date, String description, String destination, String transportMode) {
-        super(amount, date, description);
-        this.destination = destination;
-        this.transportMode = transportMode;
-    }
-
-    @Override
-    public String getType() {
-        return "Travel";
-    }
-
-    public String getDestination() {
-        return destination;
-    }
-
-    public String getTransportMode() {
-        return transportMode;
-    }
-}
-
-// UtilityExpense class
-class UtilityExpense extends Expense {
-    private String utilityType;
-
-    public UtilityExpense(double amount, Date date, String description, String utilityType) {
-        super(amount, date, description);
-        this.utilityType = utilityType;
-    }
-
-    @Override
-    public String getType() {
-        return "Utility";
-    }
-
-    public String getUtilityType() {
-        return utilityType;
-    }
-}
-
-// ExpenseManager class
-class ExpenseManager {
-    private ExpenseDAO expenseDAO;
-
-    public ExpenseManager() {
-        this.expenseDAO = new ExpenseDAO();
-    }
-
-    public void addExpense(Expense exp) {
-        expenseDAO.insertExpense(exp);
-    }
-
-    public ArrayList<Expense> getExpenses() {
-        return expenseDAO.getAllExpenses();
-    }
-
-    public double getTotalExpenses() {
-        double total = 0;
-        for (Expense exp : getExpenses()) {
-            total += exp.getAmount();
-        }
-        return total;
-    }
-}
-
-// ExpenseDAO class for database operations
-class ExpenseDAO {
-    private Connection connection;
-
-    public ExpenseDAO() {
-        connection = DatabaseConnection.getConnection();
-    }
-
-    public void insertExpense(Expense exp) {
-        String sql = "INSERT INTO expenses (amount, date, description, type, details) VALUES (?, ?, ?, ?, ?)";
-        try (PreparedStatement pstmt = connection.prepareStatement(sql)) {
-            pstmt.setDouble(1, exp.getAmount());
-            String dateStr = new SimpleDateFormat("yyyy-MM-dd").format(exp.getDate());
-            pstmt.setString(2, dateStr);
-            pstmt.setString(3, exp.getDescription());
-            pstmt.setString(4, exp.getType());
-            String details = "";
-            if (exp instanceof FoodExpense) {
-                details = ((FoodExpense) exp).getRestaurant();
-            } else if (exp instanceof TravelExpense) {
-                TravelExpense t = (TravelExpense) exp;
-                details = t.getDestination() + ", " + t.getTransportMode();
-            } else if (exp instanceof UtilityExpense) {
-                details = ((UtilityExpense) exp).getUtilityType();
-            }
-            pstmt.setString(5, details);
-            pstmt.executeUpdate();
-        } catch (SQLException e) {
-            e.printStackTrace();
-        }
-    }
-
-    public ArrayList<Expense> getAllExpenses() {
-        ArrayList<Expense> expenses = new ArrayList<>();
-        String sql = "SELECT * FROM expenses";
-        try (Statement stmt = connection.createStatement();
-                ResultSet rs = stmt.executeQuery(sql)) {
-            SimpleDateFormat sdf = new SimpleDateFormat("yyyy-MM-dd");
-            while (rs.next()) {
-                double amount = rs.getDouble("amount");
-                String dateStr = rs.getString("date");
-                Date date = sdf.parse(dateStr);
-                String description = rs.getString("description");
-                String type = rs.getString("type");
-                String details = rs.getString("details");
-                switch (type) {
-                    case "Food":
-                        expenses.add(new FoodExpense(amount, date, description, details));
-                        break;
-                    case "Travel":
-                        String[] td = details.split(", ");
-                        expenses.add(new TravelExpense(amount, date, description, td[0], td[1]));
-                        break;
-                    case "Utility":
-                        expenses.add(new UtilityExpense(amount, date, description, details));
-                        break;
-                }
-            }
-        } catch (SQLException e) {
-            e.printStackTrace();
-        } catch (ParseException e) {
-            e.printStackTrace();
-        }
-        return expenses;
-    }
-}
-
-// DatabaseConnection class for JDBC connection
-class DatabaseConnection {
-    private static final String URL = "jdbc:sqlite:expense_tracker.db";
-
-    public static Connection getConnection() {
-        try {
-            Class.forName("org.sqlite.JDBC");
-            Connection conn = DriverManager.getConnection(URL);
-            try (Statement stmt = conn.createStatement()) {
-                stmt.execute("PRAGMA foreign_keys = ON");
-            }
-            return conn;
-        } catch (Exception e) {
-            e.printStackTrace();
-            return null;
-        }
-    }
-}
-
-// Main GUI class
 public class ExpenseTrackerGUI extends JFrame {
     private ExpenseManager manager;
     private JTextField amountField;
